move queue to common
